Move impl ToTokens of semantics::type_ref
