+++ espr/src/semantics/type_ref.rs
@@ -1,8 +1,5 @@
 use super::{namespace::*, scope::*, *};
 use crate::ast;
-use inflector::Inflector;
-use proc_macro2::TokenStream;
-use quote::*;
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct SimpleType(pub ast::types::SimpleType);
@@ -15,21 +12,6 @@ impl Legalize for SimpleType {
         input: &Self::Input,
     ) -> Result<Self, SemanticError> {
         Ok(SimpleType(input.clone()))
-    }
-}
-
-impl ToTokens for SimpleType {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        use ast::types::SimpleType::*;
-        match self.0 {
-            Number => tokens.append(format_ident!("f64")),
-            Real => tokens.append(format_ident!("f64")),
-            Integer => tokens.append(format_ident!("i64")),
-            Logical => tokens.append_all(quote! { Logical }),
-            Boolen => tokens.append(format_ident!("bool")),
-            String_ { .. } => tokens.append(format_ident!("String")),
-            Binary { .. } => unimplemented!("Binary type is not supported yet"),
-        }
     }
 }
 
@@ -141,34 +123,5 @@ impl Legalize for TypeRef {
             }
             _ => todo!(),
         })
-    }
-}
-
-impl ToTokens for TypeRef {
-    fn to_tokens(&self, tokens: &mut TokenStream) {
-        use TypeRef::*;
-        match self {
-            SimpleType(ty) => ty.to_tokens(tokens),
-            Named { name, .. } => {
-                let name = format_ident!("{}", name.to_pascal_case());
-                tokens.append_all(quote! { #name });
-            }
-            Entity {
-                name,
-                has_supertype_decl,
-                ..
-            } => {
-                if *has_supertype_decl {
-                    let name = format_ident!("{}Any", name.to_pascal_case());
-                    tokens.append_all(quote! { Box<dyn #name> });
-                } else {
-                    let name = format_ident!("{}", name.to_pascal_case());
-                    tokens.append_all(quote! { #name });
-                }
-            }
-            Set { base, .. } | List { base, .. } => {
-                tokens.append_all(quote! { Vec<#base> });
-            }
-        }
     }
 }
